requirements.txt
numpy==1.26.4
pandas==2.2.3
pyyaml==6.0.2
pytest==8.3.4
scikit-learn==1.8.0
kagglehub==1.0.0
matplotlib==3.9.2

